CORRECTIONS de potentielles erreurs

+++ include/ft.h
@@ -6,7 +6,7 @@
 /*   By: thleger <thleger@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/23 15:14:34 by thleger           #+#    #+#             */
-/*   Updated: 2018/07/23 17:40:47 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/07/25 18:06:09 by thleger          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@
 #include <unistd.h>
 #include <stdlib.h>
 
-void	ft_putchar(char c);
-void	ft_putstr(char *str);
-int		ft_strlen(char *str);
-void	ft_destroy(char *str);
-int		ft_atoi(char *str);
+void					ft_putchar(char c);
+void					ft_putstr(char *str);
+unsigned long long int	ft_strlen(char *str);
+void					ft_destroy(char *str);
+unsigned long long int	ft_atoi(char *str);
 
 #endif
