+++ libft/ft_atoi.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cseguier <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: cseguier <cseguier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/08 12:02:50 by cseguier          #+#    #+#             */
-/*   Updated: 2018/07/17 10:08:53 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/07/25 18:05:42 by thleger          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(char *str)
+unsigned long long int	ft_atoi(char *str)
 {
-	int	i;
-	int	nb;
-	int sign;
+	int			i;
+	unsigned long long int	nb;
+	int 		sign;
 
 	i = 0;
 	nb = 0;
@@ -31,7 +31,7 @@ int	ft_atoi(char *str)
 	while (str[i] >= '0' && str[i] <= '9')
 	{
 		nb *= 10;
-		nb += (int)str[i] - '0';
+		nb += str[i] - '0';
 		i++;
 	}
 	if (sign)
